+++ PomocnikDozorowy/MainWindow.xaml.cs
@@ -20,5 +20,15 @@ namespace PomocnikDozorowy
         {
             InitializeComponent();
         }
+        /// <summary>
+        /// Wyświetla informacje o programie.
+        /// </summary>
+        private void InfoProgr_Click(object sender, RoutedEventArgs e)
+        {
+            MessageBox.Show("Pomocnik Dozorowy\n" +
+                "            Program dzięki któremu możesz trzymać informacje\n" +
+                "            o UTB w jednym miejscu!\n\n" +
+                "            © 2024 Pomocnik Dozorowy Team v0.0.0", "Informacje o programie", MessageBoxButton.OK, MessageBoxImage.Information);
+        }
     }
 }